+++ api/src/middleware/errorHandler.ts
@@ -4,6 +4,8 @@ import { ErrorRequestHandler } from "express";
 // Create error type
 // Error middleware function
 const errorHandler: ErrorRequestHandler = (err, req, res, next) => {
+  console.log("errorHandler running...");
+  console.log("err:", err);
   const { status, code, errors } = err;
   res.status(code).json({ status, code, errors });
 };